Main Header and Map have been updated for Redux
