--- app/components/Header/messages.js
@@ -1,13 +1,0 @@
-/*
- * HomePage Messages
- *
- * This contains all the text for the HomePage component.
- */
-import { defineMessages } from 'react-intl';
-
-export default defineMessages({
-  appName: {
-    id: 'app.components.Header.appName',
-    defaultMessage: 'venu',
-  },
-});
