+++ app/containers/App/constants.js
@@ -16,4 +16,4 @@ export const CHANGE_USER_LOCATION = 'venu/App/CHANGE_USER_LOCATION';
 export const CHANGE_USER_INTERESTS = 'venu/App/CHANGE_USER_INTERESTS';
 export const CHANGE_PARKING_LOCATION = 'venu/App/CHANGE_PARKING_LOCATION';
 export const SETUP_GEOLOCATION = 'venu/App/SETUP_GEOLOCATION';
-
+export const CHANGE_MAP_MODE = 'venu/App/CHANGE_MAP_MODE';
