+++ app/containers/App/selectors.js
@@ -6,14 +6,51 @@ import { createSelector } from 'reselect';
 
 const selectGlobal = (state) => state.get('global');
 
+// User method
 const makeSelectUser = () => createSelector(
   selectGlobal,
   (globalState) => globalState.get('user')
 );
 
+// Exhibits method
+const makeSelectExhibits = () => createSelector(
+  selectGlobal,
+  (globalState) => globalState.get('exhibits')
+);
+
+// Facilities method
+const makeSelectFacilities = () => createSelector(
+  selectGlobal,
+  (globalState) => globalState.get('facilities')
+);
+
+// Onboarding validation method
 const makeSelectOnboardingValidation = () => createSelector(
   selectGlobal,
   (onboardingState) => onboardingState.get('validation')
+);
+
+// MapMode method
+const makeSelectMapMode = () => createSelector(
+  selectGlobal,
+  (globalState) => globalState.get('mapMode')
+);
+
+// Current Place method
+const makeSelectCurrentPlace = () => createSelector(
+  selectGlobal,
+  (globalState) => globalState.get('currentPlace')
+);
+
+// Venu Map Props method
+const makeSelectVenuMap = () => createSelector(
+  selectGlobal,
+  (globalState) => globalState.get('venuMap')
+);
+
+const makeSelectDetailViewActions = () => createSelector(
+  selectGlobal,
+  (globalState) => globalState.get('detailViewActions')
 );
 
 // makeSelectLocationState expects a plain JS object for the routing state
@@ -36,6 +73,12 @@ const makeSelectLocationState = () => {
 export {
   selectGlobal,
   makeSelectUser,
+  makeSelectVenuMap,
+  makeSelectMapMode,
+  makeSelectExhibits,
+  makeSelectFacilities,
+  makeSelectCurrentPlace,
   makeSelectLocationState,
+  makeSelectDetailViewActions,
   makeSelectOnboardingValidation,
 };
