+++ app/containers/Main/Header/styles.js
